+++ include/nc_runtime/nc_polyline.h
@@ -6,14 +6,24 @@
 class NcPolyline : public NcObject
 {
 public:
-    NcPolyline(const nc::Vector2* points, int count, bool closed);
-    ~NcPolyline();
+    static forceinline NcPolyline makeWithPointsNoCopy(const nc::Vector2* points, int count, bool closed)
+    {
+        NcPolyline o;
+        o.m_points = (nc::Vector2*)points;
+        o.m_pointCount = count;
+        o.m_closed = closed;
+        return o;
+    }
 
     static sp<NcPolyline> alloc(const nc::Vector2* points, int count, bool closed)
     {
-        sp<NcPolyline> o = sp<NcPolyline>::alloc(points, count, closed);
+        sp<NcPolyline> o = NcObject::alloc<NcPolyline>(points, count, closed);
         return o;
     }
+
+    NcPolyline() : m_closed(false), m_shouldFree(false), m_pointCount(0), m_points(nullptr) {}
+    NcPolyline(const nc::Vector2* points, int count, bool closed);
+    ~NcPolyline();
 
     forceinline bool isClosed() { return m_closed; }
     forceinline int pointCount() { return m_pointCount; }
@@ -25,6 +35,7 @@ public:
 
 protected:
     bool m_closed;
+    bool m_shouldFree;
     int m_pointCount;
     nc::Vector2* m_points;
 };