+++ test/nc_data_unittest.cpp
@@ -10,4 +10,13 @@ TEST(NcData, basic)
     o = NcData::allocWithBytes("top", 4);
     EXPECT_EQ(o->length(), 4);
     EXPECT_STREQ((char*)o->bytes(), "top");
+
+    // no copy
+    {
+        const char* bytes = "hello";
+        auto o = NcData::makeWithBytesNoCopy(bytes, 5);
+        EXPECT_EQ(o.length(), 5);
+        EXPECT_TRUE(memcmp(o.bytes(), "hello", 5) == 0);
+        EXPECT_EQ(o.bytes(), bytes);
+    }
 }